klasse fjernet og tilføjet i servlet
