--- src/main/java/dat/backend/model/entities/FoodAndDrinks.java
@@ -1,17 +1,0 @@
-package dat.backend.model.entities;
-
-import java.util.List;
-
-public class FoodAndDrinks {
-
-    private String food;
-
-    private int kcal;
-
-    final static List<FoodAndDrinks> foodObjects = (new FoodAndDrinks("Pizza", 600));
-
-    public FoodAndDrinks(String food, int kcal) {
-        this.food = food;
-        this.kcal = kcal;
-    }
-}
